+++ R/spdesign-package.R
@@ -1,0 +1,10 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+#' @import cli
+#' @import stringr
+#' @importFrom stats runif qnorm as.formula model.matrix
+#' @importFrom utils modifyList
+## usethis namespace: end
+NULL