+++ NEWS.md
@@ -1,2 +1,9 @@
-## spdesignR v0.0.1
+# spdesign v0.0.2
+* New function ´probabilities()´ will now return the choice probabilities by choice task. 
+* Suppress warnings when calculating the correlation between the blocking column and the attributes to avoid warning when calculating correlation with respect to a constant.
+* After a number of candidates without improvement try a new design candidate when using the RSC algorithm
+* Updated package load message
+* Fixed roxygen @docType issue
+
+# spdesign v0.0.1
 * This is the first working version of the `spdesign` package that is able to create simple efficient designs for the MNL model.